+++ events/guildDelete.js
@@ -1,0 +1,3 @@
+module.exports = async (bot, guild) => {
+    bot.guildsettings.delete(guild.id);
+}