+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). All dates follow ISO format (year-month-day).
 
 ## [Unreleased]
+### Added
+- Per-guild configurations
+- New prefix command
 ### Changed
 - Moved global_config file to root folder
 - Miscellaneous method changes